Time: 2 ms (39.7%), Space: 8.5 MB (53.75%) - LeetHub

+++ 0007-reverse-integer/0007-reverse-integer.cpp
@@ -10,6 +10,6 @@ public:
             }
             x/=10;
         }
-        return num;
+        return (int)num;
     }
 };
